requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain-core==0.3.75
    # via
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==0.6.7
    # via -r requirements.in
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.6
    # via langgraph
langsmith==0.4.27
    # via langchain-core
markupsafe==3.0.2
    # via jinja2
narwhals==2.4.0
    # via altair
numpy==2.3.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.2
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydantic==2.11.7
    # via
    #   langchain-core
    #   langgraph
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pydub==0.25.1
    # via streamlit-audiorecorder
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
streamlit==1.49.1
    # via
    #   -r requirements.in
    #   streamlit-audiorecorder
streamlit-audiorecorder==0.0.6
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xxhash==3.5.0
    # via langgraph
zstandard==0.24.0
    # via langsmith
